test: refactor tests into subfolders
